% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_ci.R
\name{add_spectra_ci}
\alias{add_spectra_ci}
\title{Add spectra confidence interval ribbon to plot}
\usage{
add_spectra_ci(spectra_ci, ...)
}
\arguments{
\item{spectra_ci}{Spectra confidence limits object}

\item{...}{Additional arguments to `polygon`}
}
\description{
Add spectra confidence interval ribbon to plot
}
