% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prior.R
\name{check_prior}
\alias{check_prior}
\title{Check prior for validity}
\usage{
check_prior(prior, n_test = 500, error = FALSE, progress = TRUE)
}
\arguments{
\item{prior}{BayesianTools prior object}

\item{n_test}{Number of samples to test}

\item{error}{Logical. If `TRUE`, throw an error if priors are invalid. If 
`FALSE`, throw a warning and store the invalid parameters.}

\item{progress}{Logical. If `TRUE`, print a progress bar.}
}
\description{
Check prior for validity
}
