% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samples.R
\name{col_densities}
\alias{col_densities}
\title{Compute column kernel densities}
\usage{
col_densities(mat, alpha = 0.05, ...)
}
\arguments{
\item{mat}{Matrix on whose columns to compute densities}

\item{alpha}{Confidence level for trimming. Default = 0.05}
}
\description{
Compute column kernel densities
}
