% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specfile.R
\name{combine_logical}
\alias{combine_logical}
\title{Combine two logical vectors}
\usage{
combine_logical(x, y, sep = ".", ...)
}
\arguments{
\item{x}{Named logical vector}

\item{y}{Named logical vector}
}
\value{
Logical vector `x & y`, with concatenated names
}
\description{
Is identical to `&`, except that it also concatenates the names of the 
vectors.
}
