% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_functions.R
\name{convert_ed_params}
\alias{convert_ed_params}
\title{Convert raw parameter values to ED2 input format}
\usage{
convert_ed_params(params, vis = 400:700, nir = 701:1300,
  fix_allom2 = TRUE)
}
\arguments{
\item{params}{(Numeric) Long parameter vector, as sampled by
`BayesianTools`.}

\item{vis}{(Numeric) Visible wavelength range, in nm (default = `400:700`)}

\item{nir}{(Numeric) Near-infrared wavelength range, in nm (default
= `701:1300`)}

\item{fix_allom2}{(Logical) Whether or not to fix the second
allometry coefficient. (Default = `TRUE`)}
}
\description{
First, convert PROSPECT parameter values to spectra and aggregate
up to visible (VIS) and near-infrared (NIR) range. Second, if
`fix_allom2`, retrieve the allometry mean and add it to the sample list.
}
\author{
Alexey Shiklomanov
}
