% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prior.R
\name{create_prior}
\alias{create_prior}
\alias{create_prior_sampler}
\alias{create_prior_density}
\title{Create prior functions}
\usage{
create_prior(
  fix_allom2 = TRUE,
  heteroskedastic = TRUE,
  verbose = TRUE,
  nsite = 1,
  limits = FALSE,
  best = TRUE,
  param_names = NULL,
  site_specific_var = FALSE
)

create_prior_sampler(
  fix_allom2 = TRUE,
  heteroskedastic = TRUE,
  nsite = 1,
  site_specific_var = FALSE,
  param_names = NULL
)

create_prior_density(
  fix_allom2 = TRUE,
  heteroskedastic = TRUE,
  verbose = TRUE,
  param_names = NULL,
  site_specific_var = FALSE
)
}
\arguments{
\item{fix_allom2}{Logical. If `TRUE`, fix allometry exponent to prior mean. 
If `FALSE`, use full multivariate prior.}

\item{heteroskedastic}{Logical. If `TRUE`, use heteroskedastic error model 
(with residual slope and intercept). If `FALSE`, use scalar residual.}

\item{verbose}{Logical. If `TRUE`, print verbose error messages for invalid 
prior densities}
}
\description{
Create prior functions
}
