% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_allometry.R
\name{dbh2h}
\alias{dbh2h}
\title{Given DBH and PFT identity, calculate cohort height}
\usage{
dbh2h(dbh, pft, b1Ht = default_b1Ht, b2Ht = default_b2Ht)
}
\arguments{
\item{dbh}{(Numeric) Vector of cohort DBH values (must be same
length as `pft`)}

\item{pft}{(Integer) Vector of cohort PFT identities (must be same
length as `dbh`)}

\item{b1Ht}{(Numeric) PFT-specific height allometry intercept
(default defined in function)}

\item{b2Ht}{(Numeric) PFT-specific
height allometry slope (default defined in function)}
}
\value{
Numeric vector of cohort heights
}
\description{
Given DBH and PFT identity, calculate cohort height
}
\author{
Alexey Shiklomanov
}
