% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed_pft_mapping.R
\docType{data}
\name{ed_pft_mapping}
\alias{ed_pft_mapping}
\title{Mapping to ED PFTs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
ed_pft_mapping
}
\description{
Mapping to ED PFTs
}
\keyword{datasets}
