% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_date.R
\name{file_date}
\alias{file_date}
\title{Extract date from an ED output file name}
\usage{
file_date(fname)
}
\arguments{
\item{fname}{File name}
}
\value{
Date
}
\description{
Extract date from an ED output file name
}
