% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_functions.R
\name{filter_samples}
\alias{filter_samples}
\title{Remove illegal (e.g. less than zero) values, and optionally remove
duplicate rows}
\usage{
filter_samples(samples_raw, dedup = TRUE)
}
\arguments{
\item{samples_raw}{(Numeric matrix) Matrix of sample values (MCMC
iteration rows x parameter columns)}

\item{dedup}{(Logical) If `TRUE`, remove duplicate rows}
}
\value{
Numeric matrix of processed samples
}
\description{
Remove illegal (e.g. less than zero) values, and optionally remove
duplicate rows
}
\author{
Alexey Shiklomanov
}
