% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_LAI.R
\name{fraction_LAI}
\alias{fraction_LAI}
\title{Relative distribution of LAI for a specific PFT}
\usage{
fraction_LAI(h5file, ispatch = TRUE, PFTselect = 17,
  alpha_frac = 0.5)
}
\description{
Relative distribution of LAI for a specific PFT
}
