% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met.R
\name{get_cruncep}
\alias{get_cruncep}
\title{Download CRUNCEP data}
\usage{
get_cruncep(lat.in, lon.in, outfolder, start_run, end_run)
}
\arguments{
\item{lat.in}{Latitude coordinate}

\item{lon.in}{Longitude coordinate}

\item{outfolder}{Target directory for storing met}

\item{start_run}{Run start date}

\item{end_run}{Run end date (inclusive)}
}
\description{
Download CRUNCEP data
}
