% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edvar.R
\name{get_edvar}
\alias{get_edvar}
\title{Retrieve variable from ED history file}
\usage{
get_edvar(prefix, varname, edr_dir = "edr", hist_prefix = "history-S")
}
\arguments{
\item{prefix}{File path prefix}

\item{varname}{Name of variable in HDF5 file}

\item{edr_dir}{Name of EDR output subdirectory (default = "edr")}

\item{hist_prefix}{Prefix of history file name (default = "history-S")}
}
\description{
Retrieve variable from ED history file
}
