% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed_history.R
\name{get_h5vars}
\alias{get_h5vars}
\title{Read a subset of HDF5 variables into a list}
\usage{
get_h5vars(hfile, vars = NULL, except = character())
}
\arguments{
\item{hfile}{HDF5 file, as character}

\item{vars}{Character vector of variable names to read. If `NULL`, read all 
variables.}

\item{except}{Exclude these variables from the list.}
}
\value{
Named list of variables, as well as date and file name
}
\description{
Read a subset of HDF5 variables into a list
}
