% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipft.R
\name{get_ipft}
\alias{get_ipft}
\title{Get PFT index (starting from 1) of a PFT vector}
\usage{
get_ipft(pft, pft_dict = c(9, 10, 11, 6, 8))
}
\arguments{
\item{pft}{Integer vector of original (ED) PFT numbers}

\item{pft_dict}{Integer vector describing the order of PFTs (first
PFT is index 1, second PFT is index 2, etc.)}
}
\value{
Vector of PFTs
}
\description{
Thin wrapper around [base::match] with some additional error checking.
}
\author{
Alexey Shiklomanov
}
