% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_merra.R
\name{get_merra_date}
\alias{get_merra_date}
\title{Download MERRA flux variables}
\usage{
get_merra_date(date, latitude, longitude, outdir, overwrite = FALSE)
}
\description{
Download MERRA flux variables
}
\author{
Alexey Shiklomanov
}
