% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pfts.R
\name{get_pfts}
\alias{get_pfts}
\title{Extract PFTs from css file}
\usage{
get_pfts(css_file_path, delim = " ", reference = pft_lookup)
}
\arguments{
\item{css_file_path}{Path to CSS file}

\item{delim}{File delimiter. Default = " "}

\item{reference}{PFT reference table (default = pft_lookup)}
}
\description{
Extract PFTs from css file
}
