% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_ci.R
\name{get_post_ci}
\alias{get_post_ci}
\alias{get_prior_ci}
\title{Retrieve posterior or prior confidence interval for a site}
\usage{
get_post_ci(site, pda_dir, burnin, alpha = 0.05)

get_prior_ci(site, pda_dir, burnin, alpha = 0.05)
}
\arguments{
\item{site}{Site code}

\item{pda_dir}{Directory of PDA runs}

\item{burnin}{Number of samples to burnin}

\item{alpha}{P value for quantiles}
}
\description{
Retrieve posterior or prior confidence interval for a site
}
