% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_site.R
\name{get_site_prefix}
\alias{get_site_prefix}
\title{Get site prefix from site path and filename}
\usage{
get_site_prefix(site_path, which = "first")
}
\arguments{
\item{site_path}{Directory containing site css and pss files}

\item{which}{If multiple sites present, which to choose. Either "first" or "last".}
}
\value{
Site prefix, as character
}
\description{
Get site prefix from site path and filename
}
