% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat.R
\docType{data}
\name{l5bands}
\alias{l5bands}
\alias{l7bands}
\alias{l8bands}
\title{Landsat band centers, in nm}
\format{
An object of class \code{numeric} of length 6.

An object of class \code{numeric} of length 6.

An object of class \code{numeric} of length 7.
}
\usage{
l5bands

l7bands

l8bands
}
\description{
Landsat band centers, in nm
}
\keyword{datasets}
