% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_local.R
\name{load_local}
\alias{load_local}
\title{Read an `Rdata` file into a list (rather than into the global
environment)}
\usage{
load_local(file)
}
\arguments{
\item{file}{Target file name}
}
\value{
Named list containing objects in `file`.
}
\description{
Read an `Rdata` file into a list (rather than into the global
environment)
}
\examples{
\dontrun{
mylist <- load_local("file.RData")
}
}
\author{
Alexey Shiklomanov
}
