% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_observations.R
\name{load_observations}
\alias{load_observations}
\title{Load AVIRIS observations into a list}
\usage{
load_observations(
  sites,
  aviris_specfile = here::here("aviris/aviris.rds"),
  use_waves = seq(400, 1300, by = 10)
)
}
\arguments{
\item{sites}{Character vector of site tags}

\item{aviris_specfile}{Path to AVIRIS spectra H5 file. Default = "aviris/aviris.rds"}

\item{use_waves}{Wavelengths to use for observation. Default = 400 to 1300, by 10nm}
}
\description{
Load AVIRIS observations into a list
}
