% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_met.R
\name{met2ed}
\alias{met2ed}
\title{Convert meteorology to ED format}
\usage{
met2ed(rundir, full_name, latitude, longitude, overwrite = TRUE)
}
\arguments{
\item{rundir}{ED runtime directory}

\item{full_name}{Full path and prefix of meteorology files}

\item{latitude}{Latitude coordinate}

\item{longitude}{Longitude coordinate}

\item{overwrite}{Logical. If `TRUE`, overwrite existing met}
}
\description{
Convert meteorology to ED format
}
