% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.EDR.R
\name{model2netcdf.EDR}
\alias{model2netcdf.EDR}
\title{Write PEcAn standard output files}
\usage{
model2netcdf.EDR(outdir, sitelat, sitelon, start_date, par.wl = 400:2499,
  nir.wl = 2500, patches = FALSE, PFTselect = 1)
}
\description{
Write PEcAn standard output files
}
