% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_edr.R
\name{param_sub}
\alias{param_sub}
\title{Subset parameters for ith PFT}
\usage{
param_sub(i, params, pft_ends)
}
\arguments{
\item{i}{PFT index}

\item{params}{Full vector of parameter values}

\item{pft_ends}{Vector of PFT index ends}
}
\description{
Subset parameters for ith PFT
}
