% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDA_prospect.R
\name{PDA_prospect}
\alias{PDA_prospect}
\title{Run PDA on prospect}
\usage{
PDA_prospect(settings, Spectrum_leaf_data, df_PFT, wl.min, wl.max,
  use_meta.analysis = TRUE)
}
\description{
Run PDA on prospect
}
