% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pfts.R
\docType{data}
\name{pft_lookup}
\alias{pft_lookup}
\title{ED PFT lookup table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
pft_lookup
}
\description{
ED PFT lookup table
}
\keyword{datasets}
