% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patch_vertical_distribution.R
\name{plot_patch_vertical_distribution}
\alias{plot_patch_vertical_distribution}
\title{Plot patch vertical distribution}
\usage{
plot_patch_vertical_distribution(h5file, ispatch = TRUE)
}
\description{
Plot patch vertical distribution
}
