% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_sensitivity.R
\name{plot_sens}
\alias{plot_sens}
\title{Plot sensitivity confidence bands}
\usage{
plot_sens(result, obs, ...)
}
\arguments{
\item{result}{Matrix of resulting spectra}

\item{obs}{Matrix of observations}

\item{...}{Additional arguments to `plot`}
}
\description{
Plot sensitivity confidence bands
}
