% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_ci.R
\name{plot_site_spectra}
\alias{plot_site_spectra}
\title{Plot prior, posterior, and observations for a given site}
\usage{
plot_site_spectra(
  site,
  pda_dir,
  burnin,
  prior_color = alpha("blue", 0.2),
  post_color = alpha("red", 0.5),
  obs_color = "black",
  obs_lwd = 2,
  pb = NULL,
  ...
)
}
\arguments{
\item{site}{Site code}

\item{pda_dir}{Directory of PDA runs}

\item{burnin}{Number of samples to burnin}

\item{prior_color}{Color for prior band}

\item{post_color}{Color for posterior band}

\item{obs_color}{Color for observation line}

\item{obs_lwd}{Linewidth of observation line}

\item{pb}{Progress bar object (default = `NULL`)}
}
\description{
Plot prior, posterior, and observations for a given site
}
