% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_functions.R
\name{predict_site_spectra}
\alias{predict_site_spectra}
\title{Generate EDR spectra (mean, SD, and confidence envelope)
predictions for a single site}
\usage{
predict_site_spectra(params_matrix, site, nsamp = 500, dedup = FALSE,
  progress = FALSE)
}
\arguments{
\item{dedup}{(Logical) If `TRUE`, remove duplicate rows}
}
\description{
Generate EDR spectra (mean, SD, and confidence envelope)
predictions for a single site
}
\author{
Alexey Shiklomanov
}
