% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_functions.R
\name{preprocess_samples}
\alias{preprocess_samples}
\title{Convert raw `BayesianTools` samples object to an ED-compatible
parameter list}
\usage{
preprocess_samples(samples_bt, param_names, other_posteriors, nens,
  fix_allom2 = TRUE, last_n = 5000)
}
\arguments{
\item{samples_bt}{`BayesianTools` samples object, as returned by
[BayesianTools::runMCMC()].}

\item{param_names}{(Character) Vector of parameter names}

\item{other_posteriors}{Additional posterior samples to use}

\item{nens}{(Integer) Size of the resulting ensemble}

\item{fix_allom2}{(Logical) Whether or not to fix the second
allometry coefficient. (Default = `TRUE`)}

\item{last_n}{(Integer) How many samples to grab from `samples_bt`.}
}
\value{
List length `nens`, each element a named list of PFT
  parameters.
}
\description{
Convert raw `BayesianTools` samples object to an ED-compatible
parameter list
}
\author{
Alexey Shiklomanov
}
