% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samples.R
\name{prior_posterior_density}
\alias{prior_posterior_density}
\title{Prior and posterior density plots}
\usage{
prior_posterior_density(post, pri, param, ...)
}
\arguments{
\item{post}{Posterior density object}

\item{pri}{Prior density object}

\item{param}{Parameter name}
}
\description{
Prior and posterior density plots
}
