% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{rallom2}
\alias{rallom2}
\alias{dallom2}
\title{Multivariate allometry prior (varying both base and exponent)}
\usage{
rallom2()

dallom2(traits, log = TRUE)
}
\arguments{
\item{traits}{Traits list}
}
\description{
Multivariate allometry prior (varying both base and exponent)
}
