% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{rclumping}
\alias{rclumping}
\alias{dclumping}
\alias{rorient}
\alias{dorient}
\alias{rresidual}
\alias{dresidual}
\alias{rresidual2}
\alias{dresidual2}
\alias{rprospect}
\alias{dprospect}
\title{Priors on other EDR parameters}
\usage{
rclumping()

dclumping(traits, log = TRUE)

rorient()

dorient(traits, log = TRUE)

rresidual(n = 1)

dresidual(params, log = TRUE)

rresidual2(n = 1)

dresidual2(params, log = TRUE)

rprospect()

dprospect(traits, log = TRUE)
}
\arguments{
\item{traits}{Traits list}
}
\description{
Priors on other EDR parameters
}
