% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.EDR.output.R
\name{read.EDR.output}
\alias{read.EDR.output}
\title{Read all EDR outputs}
\usage{
read.EDR.output(output.path, par.wl = 400:2499, nir.wl = 2500,
  patches = FALSE, ...)
}
\description{
Read all EDR outputs
}
