% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sa.output.spectrum.R
\name{read.sa.output.spectrum}
\alias{read.sa.output.spectrum}
\title{read.sa.output.spectrum}
\usage{
read.sa.output.spectrum(traits, quantiles, pecandir, outdir, variable,
  sa.run.ids = NULL, par.wl, nir.wl)
}
\description{
read.sa.output.spectrum
}
