% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed_history.R
\name{read_ed_history}
\alias{read_ed_history}
\title{Read and organize ED history file}
\usage{
read_ed_history(historyfile)
}
\arguments{
\item{historyfile}{ED `history-S` file}
}
\value{
List containing scalar, cohort, PFT, and soil data, and other metadata. See Details.
}
\description{
Reads an ED history file into a list containing `data.frames` with similar 
themes (e.g. PFT, cohort, soil).
}
\details{
- `datetime` -- Date and time of history file

- `scalar`: Two columns -- `variable` and `value`

- `cohort`: Each row is one cohort. List columns are as follows:
 - `CB_*` -- Carbon budget. Maximum values over last 12 months + current 
 month / partial month integration
 - `RAD_PROFILE_CO` -- Radiation profile, 10 values, organized as follows: 
 (PAR, NIR, TIR) x (beam, diff) x (down, up). E.g. PAR beam down, PAR beam 
 up, PAR diff down, PAR diff up, NIR beam down, ...
 - `MORT_RATE_CO` -- Rate of mortality. Each value is a different mortality 
 type: (1) Aging, (2) Negative Carbon, (3) Treefall, (4) Cold, (5) Disturbance.

- `pft`: Each row is one PFT. List columns are as follows:
 - Length 11 -- By DBH class, in 10 cm increments. See `dbh_class`.
 - Length 8 -- Height classes. See `hgt_class`

- `soil`: Each row is a soil depth, sorted from deepest to shallowest

- `other`: Additional data that doesn't fit into other categories.

- `dbh_class` -- Vector of DBH classes

- `hgt_class` -- Vector of height classes

- `disturbance_types` -- Vector of disturbance types
}
