% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ed_history.R
\name{read_hf_var}
\alias{read_hf_var}
\title{Map-friendly way to read a variable from an HDF5 file object}
\usage{
read_hf_var(hf, variable)
}
\arguments{
\item{hf}{`H5File` object}

\item{variable}{Character string of variable to read}
}
\description{
Map-friendly way to read a variable from an HDF5 file object
}
