% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_ci.R
\name{read_spectra_history}
\alias{read_spectra_history}
\title{Read spectra history}
\usage{
read_spectra_history(site, pda_dir, burnin)
}
\arguments{
\item{site}{Site code}

\item{pda_dir}{Directory of PDA runs}

\item{burnin}{Number of samples to burnin}
}
\value{
Matrix of spectra for each iteration
}
\description{
Read spectra history
}
