% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_site_list.R
\name{refresh_site_list}
\alias{refresh_site_list}
\title{Refresh current site list based on ED runs and available spectra}
\usage{
refresh_site_list(
  pda_dir,
  aviris_specfile = here::here("aviris/aviris.rds"),
  drop_sites = "IDS05_site_1-25682",
  outfile = here::here("other_site_data/site_list")
)
}
\arguments{
\item{pda_dir}{EDR multi-site PDA directory}

\item{aviris_specfile}{Path to AVIRIS spectra H5 file. Default = "aviris/aviris.rds"}

\item{drop_sites}{Character vector of sites to drop from list.}

\item{outfile}{Target file for site list. Default = "other_site_data/site_list"}
}
\description{
Refresh current site list based on ED runs and available spectra
}
