% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pft_from_history_file.R
\name{remove_pft_from_history_file}
\alias{remove_pft_from_history_file}
\title{Remove all cohorts from one or several PFTs}
\usage{
remove_pft_from_history_file(file, file_mod, PFT2remove = c())
}
\description{
Remove all cohorts from one or several PFTs
}
