% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ci.R
\name{ribbon}
\alias{ribbon}
\title{Ribbon plot}
\usage{
ribbon(x, low, hi, ...)
}
\arguments{
\item{x}{Vector of x values}

\item{low}{Lower bounds of envelope}

\item{hi}{Upper bounds of envelope}
}
\description{
Ribbon plot
}
