% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ED_RTM.R
\name{run_ED_RTM}
\alias{run_ED_RTM}
\title{Run ED RTM}
\usage{
run_ED_RTM(rundir, soil_moisture, N, Cab, Cm, Cw, sla, b1leaf, b2leaf,
  b1Ca = NULL, b2CA = NULL, orient_factor, clumping_factor)
}
\description{
Run ED RTM
}
