% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_functions.R
\name{run_ed_site_ens}
\alias{run_ed_site_ens}
\title{Run ED for a particular site-ensemble combination}
\usage{
run_ed_site_ens(site, trait_values, outdir_prefix,
  outdir_base = "ensemble_outputs", sitedir_base = "sites",
  start_date = NULL, end_date = "2017-12-31",
  soil_info_file = file.path("other_site_data", "soil_texture.csv"),
  hetero = FALSE, fix_allom2 = TRUE, is_initial = FALSE,
  ed2in_template_file = system.file("ED2IN", package = "redr"),
  ed2in_custom = list(), ed2_exe = "ed2")
}
\arguments{
\item{site}{(Character) Site at which to run ED2}

\item{trait_values}{(List) ED parameter values}

\item{outdir_prefix}{(Character) Name of output directory. Usually
associated with an individual ensemble member (e.g. `ens_01`)}

\item{outdir_base}{(Character) Path to base output directory, which
will hold all ensemble outputs. Default = `"ensemble_outputs"`.}

\item{sitedir_base}{(Character) Path to directory containing all site information.}

\item{start_date}{(Character, formatted as date) Start date of
simulation. Defaults to June 6 in the year of the site input file.}

\item{end_date}{(Character, formatted as date) End date of
simulation. Default is 2017-12-31.}

\item{soil_info_file}{(Character) Path to file containing soil
texture information. Default is `"other_site_data/soil_texture.csv"`.}

\item{hetero}{(Logical) Whether or not the parameters use a
heteroskedastic variance model. (Default = `FALSE`)}

\item{fix_allom2}{(Logical) Whether or not to fix the second
allometry coefficient. (Default = `TRUE`)}

\item{is_initial}{(Logical) Whether the run is a bare ground run
(`TRUE`) or initial condition run (`FALSE`, default).}

\item{ed2in_template_file}{(Character) Path to ED2IN template.
Default = `"inst/ED2IN"`.}

\item{ed2in_custom}{(List) Custom ED2IN settings for ED run
(default = `list()`).}

\item{ed2_exe}{(Character) Path to ED2 executable. Passed directly
to [base::system2()], so if it is in `$PATH`, this can just be
the name of the executable. Default = `"ed2"`.}
}
\value{
List summarizing ED2 execution status.
}
\description{
Run ED for a particular site-ensemble combination
}
\author{
Alexey Shiklomanov
}
