% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_edr_date.R
\name{run_edr_date}
\alias{run_edr_date}
\title{Run EDR for a single date}
\usage{
run_edr_date(
  date,
  ed2in,
  trait_values,
  output_dir = tempdir(),
  pb = NULL,
  img_path = NULL,
  edr_exe_path = file.path("/projectnb/dietzelab/ashiklom", "ED2/EDR/build",
    "ed_2.1-opt")
)
}
\arguments{
\item{date}{Date at which to run EDR}

\item{ed2in}{ED2IN list to use as ED input}

\item{trait_values}{Named list of trait values}

\item{output_dir}{Directory where to store outputs}

\item{pb}{Optional progress bar object}

\item{img_path}{Path to Singularity container (usually a \code{.simg} file)}

\item{edr_exe_path}{If \code{img_path} is \code{NULL}, a full path to the EDR
executable. Ignored otherwise.}
}
\description{
Run EDR for a single date
}
