% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_functions.R
\name{run_edr_sample}
\alias{run_edr_sample}
\title{Run EDR at a particular site, with a particular set of parameters}
\usage{
run_edr_sample(params, isite, site_data, fix_allom2 = TRUE,
  npft = NROW(dplyr::filter(pft_lookup, !grepl("Southern_Pine",
  pft_name))), npft_param = 10, nsite = NULL, direct_sky_frac = 0.9,
  czen = 1, wavelengths = 400:2500, pb = NULL)
}
\arguments{
\item{params}{(Numeric) Long parameter vector, such as that used by
`BayesianTools`}

\item{isite}{(Integer) Site index}

\item{site_data}{(data.frame) Site cohort data, as `data.frame`.
See [PEcAn.ED2::read_css()].}

\item{npft}{(Integer) Number of PFTs. Defaults to number of rows in
[pft_lookup] that are not "Southern Pine"}

\item{npft_param}{(Integer) Number of parameters per PFT. Default = 10.}

\item{direct_sky_frac}{(Numeric) Direct sky fraction. See
[edr_r()]. Default = 0.9.}

\item{czen}{(Numeric) Cosine of solar zenith angle. See [edr_r()].
Default = 1.}

\item{wavelengths}{Numeric vector of wavelengths to use, in nm
(nwl). Default is 400:2500.}

\item{pb}{(Progress bar) Optional progress bar (default = `NULL`)}
}
\value{
Output of EDR model. See [edr_r()].
}
\description{
Run EDR at a particular site, with a particular set of parameters
}
\author{
Alexey Shiklomanov
}
