% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_edr.R
\name{run_simulation_edr}
\alias{run_simulation_edr}
\title{Run EDR simulation}
\usage{
run_simulation_edr(setup, prospect_means, prospect_covar, n_sim = 500,
  num_of_edr_params = 3)
}
\arguments{
\item{setup}{FFT setup object (see [setup_fft()])}

\item{prospect_means}{Matrix of PROSPECT parameter prior means}

\item{n_sim}{Number of simulations}

\item{num_of_edr_params}{Number of EDR params (default = 3)}
}
\value{
List of outputs. See [run_simulation_sail()].
}
\description{
Run EDR simulation
}
