% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_sail.R
\name{run_simulation_sail}
\alias{run_simulation_sail}
\title{Run SAIL simulation}
\usage{
run_simulation_sail(setup, prospect_means, prospect_covar, n_sim = 500)
}
\arguments{
\item{setup}{FFT setup object (see [setup_fft()])}

\item{prospect_means}{Matrix of PROSPECT parameter prior means}

\item{n_sim}{Number of simulations}
}
\value{
List of results (see description)
}
\description{
Returns the following:
- `refl_store` -- Stored reflectance, wavelength x simulation
- `param_store` -- Drawn parameters, param x simulation
- `ed_LAI_total` -- ED LAI summed across cohorts, and used as SAIL input
- `ed_LAI_pft` -- ED LAI for each cohort
- `ed_pft_co` -- PFT designation of each cohort
}
