% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_sensitivity.R
\name{run_site}
\alias{run_site}
\title{Perform a single EDR simulation at a single site}
\usage{
run_site(
  params,
  site,
  pda_dir,
  img_path = getOption("redr.img_path"),
  edr_exe_path = getOption("redr.edr_exe_path")
)
}
\arguments{
\item{params}{Parameter vector}

\item{site}{String identifying the site code}

\item{pda_dir}{Path to PDA run output}

\item{img_path}{Path to EDR singularity image}

\item{edr_exe_path}{Path to EDR executable}
}
\description{
Perform a single EDR simulation at a single site
}
