% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_params_edr.R
\name{sample_params_edr}
\alias{sample_params_edr}
\title{Draw parameters for EDR simulation}
\usage{
sample_params_edr(pft_vec, prospect_means, prospect_covars)
}
\arguments{
\item{pft_vec}{Vector of PFTs for which to draw parameters}

\item{prospect_means}{Matrix of PROSPECT parameter prior means}

\item{prospect_covars}{Array of PROSPECT parameter prior covariance matrices}
}
\description{
Draw parameters for EDR simulation
}
