% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_params_prospect.R
\name{sample_params_prior}
\alias{sample_params_prior}
\title{Draw parameters from prior}
\usage{
sample_params_prior(pft, prospect_means, prospect_covars)
}
\arguments{
\item{pft}{Name of PFT. Must be in `prospect_means`}

\item{prospect_means}{Matrix of PROSPECT parameter prior means}

\item{prospect_covars}{Array of PROSPECT parameter prior covariance matrices}
}
\description{
Draw parameters from prior
}
