% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_params_sail.R
\name{sample_params_sail}
\alias{sample_params_sail}
\title{Draw a random sample of PRO4SAIL parameters}
\usage{
sample_params_sail(
  pft,
  lai,
  prospect_means,
  prospect_covars,
  custom_sail = list()
)
}
\arguments{
\item{pft}{Name of PFT. Must be in `prospect_means`}

\item{lai}{Value of leaf area index}

\item{prospect_means}{Matrix of PROSPECT parameter prior means}

\item{prospect_covars}{Array of PROSPECT parameter prior covariance matrices}

\item{custom_sail}{Named list of sail parameters to replace defaults}
}
\description{
Draw a random sample of PRO4SAIL parameters
}
