% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_edr_multisite.R
\name{setup_edr_multisite}
\alias{setup_edr_multisite}
\title{Setup EDR runs for a bunch of sites}
\usage{
setup_edr_multisite(sites, pda_dir)
}
\arguments{
\item{sites}{Character vector of site names}

\item{pda_dir}{EDR PDA directory (containing all sites)}
}
\description{
Setup EDR runs for a bunch of sites
}
