% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_fft.R
\name{setup_fft}
\alias{setup_fft}
\title{Prepare an FFT plot for testing}
\usage{
setup_fft(fft_plot, meas_year, clobber = TRUE)
}
\arguments{
\item{fft_plot}{FFT plot code, as character}

\item{meas_year}{Measurement year, as integer or character}

\item{clobber}{Overwrite existing ED run if present (default = TRUE)}
}
\value{
List, containing prefix path and paths to css, pss, and site files
}
\description{
Does the following:
- Identify css, pss, and site files
- Run ED at that site
}
