% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_sensitivity.R
\name{site_sensitivity}
\alias{site_sensitivity}
\title{Perform site sensitivity analysis over a set of samples for a given site}
\usage{
site_sensitivity(samples_mat, site, pda_dir, n = 500, ...)
}
\arguments{
\item{samples_mat}{Matrix of samples. Each iteration is a row, each column is a parameter.}

\item{site}{Site to simulate}

\item{n}{Number of simulations to do. Default = 500.}
}
\description{
Perform site sensitivity analysis over a set of samples for a given site
}
