% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat.R
\name{spec2landsat}
\alias{spec2landsat}
\title{Convert full reflectance spectra to Landsat reflectance}
\usage{
spec2landsat(reflectance)
}
\arguments{
\item{reflectance}{Vector of reflectance values}
}
\value{
`tibble` containing wave
}
\description{
Convert full reflectance spectra to Landsat reflectance
}
