% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_ci.R
\name{spec_ci}
\alias{spec_ci}
\title{Draw spectra confidence interval}
\usage{
spec_ci(refl_mat, wl = 400:2500, alpha = 0.05, ...)
}
\arguments{
\item{refl_mat}{Matrix of reflectance values (wavelengths in rows)}

\item{wl}{Vector of wavelengths}

\item{alpha}{Quantile probability (e.g. 0.05 means 95\% CI)}
}
\description{
Based on empirical quantiles of reflectance spectra.
}
