% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samples.R
\name{split_params}
\alias{split_params}
\title{Split parameters column into categories}
\usage{
split_params(.data, param_col, ...)
}
\arguments{
\item{.data}{Data frame containing parameter colum}

\item{param_col}{Name of parameter column}

\item{...}{Additional arguments to [tidyr::separate]}
}
\description{
Split parameters column into categories
}
