% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_samples.R
\name{summarize_samples}
\alias{summarize_samples}
\title{Summarize samples as a tidy data frame}
\usage{
summarize_samples(samples)
}
\arguments{
\item{samples}{MCMC samples object}
}
\description{
Summarize samples as a tidy data frame
}
