% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_texture_codes.R
\docType{data}
\name{texture_codes}
\alias{texture_codes}
\title{USDA texture code conversion table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42 rows and 2 columns.
}
\usage{
texture_codes
}
\description{
USDA texture code conversion table
}
\keyword{datasets}
