% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_edr.R
\name{vec2list}
\alias{vec2list}
\title{Convert EDR parameter vector to EDR input list}
\usage{
vec2list(params, pft_vec, ...)
}
\arguments{
\item{params}{Vector of EDR parameters (see [sample_params_edr()])}
}
\description{
Convert EDR parameter vector to EDR input list
}
