% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_ci.R
\name{wcl}
\alias{wcl}
\title{Count the number of lines in a file}
\usage{
wcl(file)
}
\arguments{
\item{file}{Path to file}
}
\value{
Number of lines in file
}
\description{
Count the number of lines in a file
}
