% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_prog.R
\name{with_prog}
\alias{with_prog}
\title{Adverb to add a progress bar to a function}
\usage{
with_prog(.f, total, pb_opts = list())
}
\arguments{
\item{.f}{Function}

\item{total}{Total number of progress bar ticks}

\item{pb_opts}{Named list of other options to be passed to `progress_bar$new`}
}
\value{
Function with progress bar
}
\description{
Adverb to add a progress bar to a function
}
