% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sa.output.leafspectrum.R
\name{read.sa.output.leafspectrum}
\alias{read.sa.output.leafspectrum}
\title{read.sa.output.spectrum}
\usage{
read.sa.output.leafspectrum(traits, quantiles, pecandir, outdir, variable,
  sa.run.ids = NULL, par.wl, nir.wl, spectra_list_all)
}
\description{
read.sa.output.spectrum
}
