% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda_edr.R
\name{run_pda_edr}
\alias{run_pda_edr}
\title{Run EDR parameter data assimilation}
\usage{
run_pda_edr(setup, observed, observation_operator, prospect_means,
  prospect_covar, num_of_edr_params = 3)
}
\arguments{
\item{setup}{FFT setup object (see [setup_fft()])}

\item{observed}{Observation vector or matrix (see [PEcAnRTM::invert_bt()])}

\item{observation_operator}{Function for aligning EDR output with observation}

\item{prospect_means}{Matrix of PROSPECT parameter prior means}

\item{num_of_edr_params}{Number of EDR params (default = 3)}
}
\value{
Output of [PEcAnRTM::invert_bt()]
}
\description{
Run EDR parameter data assimilation
}
