% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation_edr.R
\name{vec2list}
\alias{vec2list}
\title{Convert EDR parameter vector to EDR input list}
\usage{
vec2list(params, pft_vec, pft_ends, ...)
}
\arguments{
\item{params}{Full vector of parameter values}

\item{pft_vec}{Vector of PFT names}

\item{pft_ends}{Vector of PFT index ends}
}
\description{
Convert EDR parameter vector to EDR input list
}
